Patch 0.1.3: Refactored header component
