+++ src/components/header/Header.js
@@ -2,13 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {makeStyles} from '@material-ui/core/styles';
 import Toolbar from '@material-ui/core/Toolbar';
-import HeaderNav from './subcomponents/HeaderNav.js';
+import HeaderNav from './HeaderNav.js';
 import { motion } from "framer-motion";
 import { GiAcorn } from "react-icons/gi";
 import { IconContext } from "react-icons";
 import theme from "../common/theme";
 import AppBar from "@material-ui/core/AppBar";
 import { Link, useLocation } from "react-router-dom";
+import HeaderLogo from "./HeaderLogo";
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -44,13 +45,7 @@ function Header(props) {
             <React.Fragment>
                 <AppBar className={classes.altRoot}>
                     <Toolbar component="nav" variant="regular" key={"header"}>
-                        <motion.div key={"headerIcon"} whileHover={{ rotate: 360, scale: 1.1, transition: {duration: 1} }}>
-                            <IconContext.Provider value={{ style: {fontSize: '35px', color: theme.palette.primary.contrastText}}}>
-                                <Link underline="none" to={'/'}>
-                                    <GiAcorn />
-                                </Link>
-                            </IconContext.Provider>
-                        </motion.div>
+                        <HeaderLogo/>
                         <div className={classes.space} key={"headerDiv"}/>
                         {sections.map((section) => (
                             <HeaderNav section={section} link={section.url} key={"link" + section.title}/>
@@ -65,13 +60,7 @@ function Header(props) {
             <React.Fragment>
                 <AppBar className={classes.root}>
                     <Toolbar component="nav" variant="regular" key={"header"}>
-                        <motion.div key={"headerIcon"} whileHover={{ rotate: 360, scale: 1.1, transition: {duration: 1} }}>
-                            <IconContext.Provider value={{ style: {fontSize: '35px', color: theme.palette.primary.main}}}>
-                                <Link underline="none" to={'/'}>
-                                    <GiAcorn />
-                                </Link>
-                            </IconContext.Provider>
-                        </motion.div>
+                        <HeaderLogo/>
                         <div className={classes.space} key={"headerDiv"}/>
                         {sections.map((section) => (
                             <HeaderNav section={section} link={section.url} key={"link" + section.title}/>
@@ -88,5 +77,4 @@ export default Header;
 
 Header.propTypes = {
     sections: PropTypes.array,
-    title: PropTypes.string,
 };
